# backend/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aster API",
        "version": "1.0.0",
        "description": "The Aster API the Android app calls. The contract is \u00a75 of docs/aster-backend-plan.md. Send `Accept: application/json` and `Authorization: Bearer <token>`: a device token from /auth/register or /auth/login, or a profile token from /profiles/{id}/unlock."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/conversations/{conversation}": {
            "delete": {
                "operationId": "conversations.destroy",
                "tags": [
                    "Conversation"
                ],
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "description": "The conversation ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/profiles/{profile}": {
            "put": {
                "operationId": "profiles.update",
                "tags": [
                    "Profile"
                ],
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "required": true,
                        "description": "The profile ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProfileResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "profiles.destroy",
                "tags": [
                    "Profile"
                ],
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "required": true,
                        "description": "The profile ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/reminders/{reminder}": {
            "put": {
                "operationId": "reminders.update",
                "tags": [
                    "Reminder"
                ],
                "parameters": [
                    {
                        "name": "reminder",
                        "in": "path",
                        "required": true,
                        "description": "The reminder ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReminderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ReminderResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ReminderResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "reminders.destroy",
                "tags": [
                    "Reminder"
                ],
                "parameters": [
                    {
                        "name": "reminder",
                        "in": "path",
                        "required": true,
                        "description": "The reminder ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/vip-contacts/{vipContact}": {
            "put": {
                "operationId": "vip-contacts.update",
                "tags": [
                    "VipContact"
                ],
                "parameters": [
                    {
                        "name": "vipContact",
                        "in": "path",
                        "required": true,
                        "description": "The vip contact ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VipContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`VipContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VipContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "vip-contacts.destroy",
                "tags": [
                    "VipContact"
                ],
                "parameters": [
                    {
                        "name": "vipContact",
                        "in": "path",
                        "required": true,
                        "description": "The vip contact ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "MessageSource": {
                "type": "string",
                "enum": [
                    "whatsapp",
                    "google_chat",
                    "sms"
                ],
                "title": "MessageSource"
            },
...
            "ProfileResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "has_pin": {
                        "type": "boolean"
                    },
                    "personality": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "web_mode": {
                        "type": "string"
                    },
                    "memory_enabled": {
                        "type": "boolean"
                    },
                    "sampling": {
                        "type": "string"
                    },
                    "max_tokens": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "auto_lock_minutes": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "color",
                    "role",
                    "has_pin",
                    "personality",
                    "web_mode",
                    "memory_enabled",
                    "sampling",
                    "max_tokens",
                    "auto_lock_minutes",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProfileResource"
            },
...
            "ReminderRequest": {
                "type": "object",
                "description": "POST needs client_uuid, title and trigger; PATCH and PUT change only what\nthey send. The controller checks that a time reminder ends up with a due_at.",
                "properties": {
                    "client_uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "trigger": {
                        "$ref": "#/components/schemas/ReminderTrigger"
                    },
                    "due_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ReminderStatus"
                    }
                },
                "required": [
                    "client_uuid",
                    "title",
                    "trigger"
                ],
                "title": "ReminderRequest"
            },
            "ReminderResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "client_uuid": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "due_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "trigger": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "client_uuid",
                    "title",
                    "due_at",
                    "trigger",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "title": "ReminderResource"
            },
            "ReminderStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "done",
                    "cancelled"
                ],
                "title": "ReminderStatus"
            },
            "ReminderTrigger": {
                "type": "string",
                "enum": [
                    "time",
                    "arrive_home"
                ],
                "title": "ReminderTrigger"
            },
            "Sampling": {
                "type": "string",
                "enum": [
                    "auto",
                    "precise",
                    "balanced",
                    "creative"
                ],
                "title": "Sampling"
            },
...
            "UpdateProfileRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40
                    },
                    "color": {
                        "type": "string",
                        "pattern": "^#[0-9A-F]{6}$"
                    },
                    "personality": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 2000
                    },
                    "web_mode": {
                        "$ref": "#/components/schemas/WebMode"
                    },
                    "memory_enabled": {
                        "type": "boolean"
                    },
                    "sampling": {
                        "$ref": "#/components/schemas/Sampling"
                    },
                    "max_tokens": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 16,
                        "maximum": 8192
                    },
                    "auto_lock_minutes": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1,
                        "maximum": 60
                    },
                    "pin": {
                        "type": "string",
                        "pattern": "^\\d{6}$"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "owner",
                            "member",
                            "restricted",
                            "guest"
                        ]
                    }
                },
                "title": "UpdateProfileRequest"
            },
            "VipContactRequest": {
                "type": "object",
                "description": "POST needs every field; PATCH and PUT change only what they send.",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 60
                    },
                    "match_key": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "sources": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MessageSource"
                        },
                        "minItems": 1,
                        "uniqueItems": true
                    },
                    "enabled": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "name",
                    "match_key",
                    "sources"
                ],
                "title": "VipContactRequest"
            },
            "VipContactResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "match_key": {
                        "type": "string"
                    },
                    "sources": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "match_key",
                    "sources",
                    "enabled",
                    "updated_at"
                ],
                "title": "VipContactResource"
            },
            "WebMode": {
                "type": "string",
                "enum": [
                    "auto",
                    "never"
                ],
                "title": "WebMode"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}